Add some basic test
Add a test that tests that libssh2_session_init and friends is called
with the expected arguments to have a starting point for making some
more useful tests.

This also show that we cannot really mock out the call to libssh2_init
as that is really global and called before main is invoked.