Lowest supported boost version is 1.77
This library requires at least boost 1.77 since that seems to be the
first version to fully support conversion between std::error_code (and
related types) and boost::system types.

We could support earlier boost versions but I cannot be bothered so
just test that we support at least boost 1.77 on (Ubuntu) Linux and
add that to the minimal README file.