Copy structured binding for use trompeloeil side effect
A structured binding is not a normal variable and cannot be used in
the trompeloeil LR_SIDE_EFFECT which uses a lambda or something
similar.

Clang correctly refuses to compile that although GCC accepts it. The
standard says it should be refused because C++ is a weird language.

Just copy it to a "normal" variable as a workaround.