Fix path conversion on Windows
On Windows .c_str() for paths will return a wchar_t so convert the
path to a standard string to do convert the path first.

This most likely in no way guarantees that this will actually create a
valid UTF-8 string but for now it will at least compile and handle
ASCII paths.